+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v37.0.0-SNAPSHOT - unreleased
 
+### 🎁 New Features
+
+* `Select` supports a new `menuWidth` prop to control the width of the dropdown.
+
 [Commit Log](https://github.com/xh/hoist-react/compare/v36.3.0...develop)
 
 ## v36.3.0 - 2020-10-07